movie-recommender/backend/requirements.txt
flask
faiss-cpu
pandas
numpy
sentence-transformers
gunicorn
